+++ rusk/tests/services/mod.rs
@@ -7,6 +7,7 @@
 pub mod contract_deployment;
 pub mod contract_stake;
 pub mod conversion;
+pub mod disable_phoenix;
 //pub mod deploy;
 pub mod event_reverts;
 pub mod fee_refund_inflation;